another broken attempt to handle the "Copy".
Using a more structured representation for the generated types
seems like a useful idea, but using unions for determining
the maximum size of the structure - is not.

Seems like some kind of "DataSizeOf" trait will be a better
idea with a single fn that returns the storage size of
the data type.