remove the hard dependency on vpp-api-transport, and delete reqrecv
it is now sitting in the vpp-api-transport crate, so use
that implementation.

Move the interactions from generated vpp-api code to vpp-api-transport

The reason is to clean up the dependencies and avoid
having code generated by vpp-api-gen depend on
vpp-api-transport for anything but examples,
because it may clash with the usage of vpp-api-transport
by the end-user code. Other than just bloat,
two different versions of vpp-api-transport linking with
a shared library make the compiler very upset.

With this change, the overall dependency graph
for the real use will look as follows:

                        vpp-api-encoding
                             ^
                             |
vpp-api-message  <---- GENERATED vpp-api ---> vpp-api-macros
     ^                 /                ^
     |                /                  \
     |               /                    \
     |            examples/tests only      \
     |           /                          \
     |          V                            \
vpp-api-transport <------------------------- end-user code

This way the "logical representation" of the messages sits
in the generated vpp-api crate, and the "transport
interaction stuff" sits in the vpp-api-transport.

I am not entirely happy with the vpp-api-transport getting
the higher level functions, but it seems the most reasonable so far,
better than replicating the code with every generated API crate.
That is unreasonable.